requirements.txt
pandas
matplotlib
pyserial
streamlit
numpy
